Use `ruby2_keywords` for the moment (#205)
Without this gem, current activeadmin version becomes incompatible with
the master branch of `arbre`. It shouldn't be an issue as long as we
update activeadmin's code to use keyword arguments everywhere, set a
minimum dependency on the next `arbre` and release both gems at the same
time, but I think it's easier to be conservative.